+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## Unreleased
 
 * Feat: Attach Isolate name to thread context (#847)
+* Fix: Fix `SentryAssetBundle` on Flutter >= 3.1 (#877)
 * Feat: Add Android thread to platform stacktraces (#853)
 * Fix: Rename auto initialize property (#857)
-* Bump: Sentry-Android to 6.0.0-beta.4 (#871)
+* Bump: Sentry-Android to 6.0.0 (#879)
 
 ## 6.6.0-alpha.2